small refactoring done + test made todo
